reverse a number
in this program we will calculte the reverse of a number. in this program we use a logic i.e. first of all we calculate the modulas of the number and then a new variable which is multiply by 10 and then add remainder.and after that we divide the number by 10. and this loop is execute untill the number become zero.