print prime numbers between two numbers
it will print all the prime number between two numbers.
the logic is that first we start a loop from start+1 and we call a function to check whether this number is prime or not and this function return a flag value if it is 1 then it is not prime otherwise yes.
then we print all the numbers which are prime.